% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accrual.R
\name{accrual}
\alias{accrual}
\title{A class to contain the accrued patient allocations,
by site, treatment/control arm and week}
\usage{
accrual(
  treatment_arm_ids = S7::class_missing,
  shared_control = S7::class_missing,
  centres_df = S7::class_missing,
  accrual_period = S7::class_missing,
  ...
)
}
\description{
A class to contain the accrued patient allocations,
by site, treatment/control arm and week
}
\section{Slots}{

\describe{
\item{\code{accrual}}{3-D array with axes site, experimental arm
and week}

\item{\code{phase_changes}}{Vector of week numbers when arms closed}
}}

