% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graphics.R
\name{accrual_arm_plot}
\alias{accrual_arm_plot}
\title{Plot single arm accrual plot}
\usage{
accrual_arm_plot(data_df, arm_colours, treatment_arms, targets, plot_id, i)
}
\arguments{
\item{data_df}{Dataframe of biomkrAccrual output.}

\item{arm_colours}{Vector of hexadecimal colours, one for each arm.}

\item{treatment_arms}{Vector of names of the treatment arms.}

\item{targets}{Vector of target names (excluding the
word \code{target}).}

\item{plot_id}{Vector of plot type names (typically "Interim" and "Accrual").}

\item{i}{Index of treatment arm to plot.}
}
\description{
Bodge, fix later
}
