% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graphics.R
\name{accrual_to_long}
\alias{accrual_to_long}
\title{Convert accrual data to long format dataframe of class
\code{accrualplotformat}.}
\usage{
accrual_to_long(accrual_df)
}
\arguments{
\item{accrual_df}{Wide format accrual data.}
}
\description{
Convert accrual data to long format dataframe of class
\code{accrualplotformat}.
}
