% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graphics.R
\name{accrual_to_target_plot_from_file}
\alias{accrual_to_target_plot_from_file}
\title{Plot predicted week at which recruitment reaches specified targets
for each arm from file containing a json consisting of a list
of matrices of arm recruitment by week, one for each arm.}
\usage{
accrual_to_target_plot_from_file(
  file_prefix = "arm_accrual",
  plot_prefix = "accrual-arm-target-week-from-file",
  run_time = "26-02-08_14-15-05",
  output_path = "../biomkrAccrual_output_data/",
  figs_path = paste0(output_path, "figures/"),
  target_treatment = NA_integer_,
  target_control = NA_integer_,
  arm_colours = NULL
)
}
\arguments{
\item{file_prefix}{Consistent beginning of filenames holding
arm closure data. Defaults to \code{accrual}.}

\item{plot_prefix}{Prefix for file name to identify plot type.
Defaults to \code{accrual_plot}.}

\item{run_time}{Specify a particular instance of \code{biomkrAccrual()}
execution using a date-time format \code{yyyy-mm-dd-hh-mm-ss}.
Used to select which files will be summarised.}

\item{output_path}{Directory where the output files from the
\code{biomkrAccrual()} instance are located.}

\item{figs_path}{Folder where figures generated during execution
will be stored; defaults to the \code{figures} subdirectory in
\code{output_path}.}

\item{target_treatment}{Recruitment target for treatment arms.}

\item{target_control}{Recruitment target for control arms.}

\item{arm_colours}{Vector of hexadecimal colours, one for each arm.
If only one value is supplied, it will be used for all arms. If no
value is supplied, a colourblind-friendly palette will be used.}
}
\description{
Plot predicted week at which recruitment reaches specified targets
for each arm from file containing a json consisting of a list
of matrices of arm recruitment by week, one for each arm.
}
