% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accrual.R
\name{accrue_week}
\alias{accrue_week}
\title{Generate accrual for a week and assign it to the
accrual object. Increments the property "week",
which holds the next week number to accrue.}
\usage{
accrue_week(obj, ...)
}
\arguments{
\item{obj}{An object of class "accrual"}
}
\description{
Generate accrual for a week and assign it to the
accrual object. Increments the property "week",
which holds the next week number to accrue.
}
