% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-batches.R
\name{biomkrAccrualSim}
\alias{biomkrAccrualSim}
\title{Run a number of batches of recruitment prediction and
collect summary statistics on arm closures, and final
recruitment totals for all experimental and control arms.
Seeds for each run are generated using the L'Ecuyer
pseudo-random number generator.}
\usage{
biomkrAccrualSim(
  n = 100,
  shared_control = TRUE,
  target_times = c(6, 12),
  precision = NULL,
  var_lambda = NULL,
  control_ratio = c(1, 1),
  centres_file = "centres.csv",
  prop_file = "proportions.csv",
  target_file = "targets.csv",
  arms_file = "arms.json",
  data_path = "extdata/",
  output_path = "../biomkrAccrual_output_data/",
  figs_path = paste0(output_path, "figures/"),
  fixed_centre_starts = TRUE,
  fixed_site_rates = FALSE,
  fixed_region_prevalences = FALSE,
  quietly = FALSE,
  seed = 123456
)
}
\arguments{
\item{n}{Number of instances to run (defaults to 100)}

\item{shared_control}{TRUE if all treatment arms share the
same control arm; FALSE if each treatment arm has its own
control. Defaults to TRUE.}

\item{target_times}{Vector of timings of interim and final
recruitment assessments, in months.}

\item{precision}{For the Dirichlet model of biomarker prevalences,
variability decreases as precision increases. Defaults to NULL.}

\item{var_lambda}{Variance estimate for site recruitment rates.
Defaults to NULL.}

\item{control_ratio}{Ratio of patient allocation to treatment arm
versus control for all active arms; defaults to c(1, 1).}

\item{centres_file}{Name of CSV file with information about
each recruitment centre; this should have columns "site",
"start_month", "mean_rate", "region" and optionally "site_cap"
if recruitment is capped per site. Defaults to \code{centres.csv}.}

\item{prop_file}{Name of CSV file with expected biomarker prevalences
for each region; this should have one column "category", naming
the biomarkers or biomarker combinations, and one column per
region. Defaults to \code{proportions.csv}.}

\item{target_file}{Name of CSV file with target recruitment for each
arm at each interim analysis time and at final recruitment. This
should have a column "target" with the arm name; this must match
the arm name as specified in in \code{arms_file}. Control targets are
not included as they can be deduced from the control_ratio and
shared_control arguments. It may then have one or more columns for
interim targets, and must have a "final" column for the final
recruitment target.}

\item{arms_file}{Name of JSON file describing which recruitment
arms (defined by biomarkers) recruit to which treatment arms.
Defaults to \code{arms_json}.}

\item{data_path}{Folder where \code{centres_file}, \code{prop_file} and
\code{arms_file} are located. Defaults to the location of the package
example data in the package installation; this should be changed.}

\item{output_path}{Folder where data generated during execution
will be stored; defaults to \verb{../biomkrAccrual_output_data/}.}

\item{figs_path}{Folder where figures generated during execution
will be stored; defaults to the \code{figures} subdirectory in
\code{output_path}.}

\item{fixed_centre_starts}{TRUE if centres are assumed to start
exactly when planned; FALSE if some randomisation should be added.}

\item{fixed_site_rates}{TRUE if centre recruitment rates should
be treated as exact; FALSE if they should be drawn from a gamma
distribution with a mean of the specified rate.}

\item{fixed_region_prevalences}{TRUE if biomarker prevalences
should be considered to be identical for all sites within a
region; FALSE if they should be drawn from a Dirichlet distribution
with a mean of the specified prevalence.}

\item{quietly}{Defaults to FALSE, which displays the output from
each run. Set to TRUE to generate data and figures without displaying
them.}

\item{seed}{Seed to be used to generate the seeds for each simulation.}
}
\value{
Dataframe of site closing times

Dataframe of experimental arm totals
}
\description{
Run a number of batches of recruitment prediction and
collect summary statistics on arm closures, and final
recruitment totals for all experimental and control arms.
Seeds for each run are generated using the L'Ecuyer
pseudo-random number generator.
}
