% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graphics.R
\name{bma_colours}
\alias{bma_colours}
\title{Custom colours (colourblind friendly) -
omits colours nearest black and white}
\usage{
bma_colours(palette = "viridis", no_colours = 4)
}
\arguments{
\item{palette}{Defaults to "viridis"}

\item{no_colours}{Defaults to 4}
}
\description{
Custom colours (colourblind friendly) -
omits colours nearest black and white
}
