% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accrual.R
\name{do_choose_cap}
\alias{do_choose_cap}
\title{Select arms to cap for single site, or sites to cap for single arm}
\usage{
do_choose_cap(population, captotal)
}
\arguments{
\item{population}{Vector of allocations of patients this week}

\item{captotal}{Value of cap for site or arm (as appropriate)}
}
\value{
Vector of arms to cap (can include multiples of the same arm)
}
\description{
Select arms to cap for single site, or sites to cap for single arm
}
