% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{do_clean_centres}
\alias{do_clean_centres}
\title{Convert centres_df to more useable format}
\usage{
do_clean_centres(centres_df)
}
\arguments{
\item{centres_df}{Dataframe with columns site, start_month, mean_rate,
prevalence_set and site_cap}
}
\value{
Dataframe with start_week, ordered by start_month and site
}
\description{
Convert centres_df to more useable format
}
