% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spine.R
\name{do_sample}
\alias{do_sample}
\title{Driver for single instance of sampling}
\usage{
do_sample(
  prop_params_df,
  arms_ls,
  trial_structure_instance,
  target_arm_size,
  margin,
  av_site_rate_month,
  accrual_period,
  goals_df,
  ctrl_ratio,
  data_path,
  centre_start_file,
  fixed_centre_starts,
  fixed_site_rates
)
}
\description{
Driver for single instance of sampling
}
