% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{do_sensitivity}
\alias{do_sensitivity}
\title{Sensitivity analysis for simplest case; fixed site rates, simultaneous centre
starts and no site capping.}
\usage{
do_sensitivity(target_arm_size, site_rates, no_centres, figs_path)
}
\arguments{
\item{target_arm_size}{Number of patients to be recruited.}

\item{site_rates}{Vector of site rates, one per site; or scalar if all
identical.}

\item{no_centres}{Number of sites.}
}
\description{
Sensitivity analysis for simplest case; fixed site rates, simultaneous centre
starts and no site capping.
}
