% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{enable-for-old-R}
\alias{enable-for-old-R}
\title{Enable usage of <S7_object>@name in package code even for
older versions of R.}
\description{
Enable usage of <S7_object>@name in package code even for
older versions of R.
}
