% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-utils.R
\name{expand_targets}
\alias{expand_targets}
\title{Expand target_df to include control arm(s)}
\usage{
expand_targets(target_df, shared_control = TRUE, control_ratio = c(1, 1))
}
\arguments{
\item{target_df}{Dataframe of recruitment targets, as read from target file.}

\item{shared_control}{One shared control arm or one per experimental arm
(logical, defaults to TRUE)?}

\item{control_ratio}{Ratio of patient allocation to treatment arm
versus control for all active arms; defaults to c(1, 1).}
}
\description{
Expand target_df to include control arm(s)
}
