% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{extend_week}
\alias{extend_week}
\title{Extend predefined array along week axis}
\usage{
extend_week(accrual, week_acc)
}
\arguments{
\item{accrual}{accrual array}

\item{week_acc}{accrual for week}
}
\description{
Extend predefined array along week axis
}
