% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graphics.R
\name{get_arm_closures}
\alias{get_arm_closures}
\title{Generate arm closure summaries for batches}
\usage{
get_arm_closures(
  file_prefix = "closures",
  run_time = "2024-08-07-18-35-09",
  output_path = "../biomkrAccrual_output_data/"
)
}
\arguments{
\item{file_prefix}{Consistent beginning of filenames holding
arm closure data. Defaults to \code{closures}.}

\item{run_time}{Specify a particular instance of \code{spine()}
execution using a date-time format \code{yyyy-mm-dd-hh-mm-ss}.
Used to select which files will be summarised.}

\item{output_path}{Directory where the input files are located
and the output files will be written.}
}
\description{
Generate arm closure summaries for batches
}
