% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{get_array_prevalence}
\alias{get_array_prevalence}
\title{Make array with the prevalences by treatment arm, recruitment arm and
prevalence set}
\usage{
get_array_prevalence(arm_structure_mx, recruit_arm_prevalence)
}
\description{
Make array with the prevalences by treatment arm, recruitment arm and
prevalence set
}
