% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{get_matrix_prevalence}
\alias{get_matrix_prevalence}
\title{Make matrix with the prevalences by treatment arm and recruitment arm}
\usage{
get_matrix_prevalence(arm_structure_mx, recruit_arm_prevalence)
}
\description{
Make matrix with the prevalences by treatment arm and recruitment arm
}
