% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{get_matrix_struct}
\alias{get_matrix_struct}
\title{Converts trial structure and prevalence information into matrix form}
\usage{
get_matrix_struct(arms_ls, recruit_arm_prevalence)
}
\description{
Converts trial structure and prevalence information into matrix form
}
