% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{get_recruit_arm_prevalence}
\alias{get_recruit_arm_prevalence}
\title{Dirichet regression model for biomarker prevalence}
\usage{
get_recruit_arm_prevalence(
  props_df,
  centres_df,
  precision,
  fixed_region_prevalences
)
}
\arguments{
\item{props_df}{Dataframe with one row per biomarker.  Has one column
\code{category}, containing names of the biomarkers, plus one column for
each region, containing the expected biomarker prevalences for the regions.}

\item{centres_df}{Dataframe with one row per site, including one column
\code{region}, containing the index number for the region for each site.
Indices are assumed to be in the same order as the columns in \code{props_df}.}

\item{precision}{Variability decreases as precision increases.}
}
\value{
Matrix of prevalences with one column per site and one
row per biomarker; each column sums to 1.
}
\description{
For each site, draws from the Dirichlet distribution using the
expected prevalences for the region.
expected prevalences by region
}
