% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{get_weeks}
\alias{get_weeks}
\title{Get no. weeks from months
Currently 4 week month, fix later using lubridate}
\usage{
get_weeks(months)
}
\arguments{
\item{months}{Duration in months}
}
\value{
Duration in weeks
}
\description{
Get no. weeks from months
Currently 4 week month, fix later using lubridate
}
