% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-batches.R
\name{getseeds}
\alias{getseeds}
\title{Set seeds for multiple simulations, using a new L'Ecuyer
stream for each one.}
\usage{
getseeds(n = 5, seed = NULL)
}
\arguments{
\item{n}{No. runs}

\item{seed}{Starting seed}
}
\value{
List of N sets of 7 seeds
}
\description{
Set seeds for multiple simulations, using a new L'Ecuyer
stream for each one.
}
