% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{ggclosures}
\alias{ggclosures}
\title{Display arm closure summaries}
\usage{
ggclosures(
  file_prefix = "arm_closures_24-04-29_",
  output_path = "output_data/"
)
}
\arguments{
\item{file_prefix}{Consistent beginning of filename}

\item{output_path}{Directory where output .csvs are written}
}
\description{
Display arm closure summaries
}
