% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graphics.R
\name{ggscatterError}
\alias{ggscatterError}
\title{Plot a scatter plot with error bars for time to recruit against prevalence}
\usage{
ggscatterError(prevalences, e_time, v_time)
}
\arguments{
\item{prevalences}{Vector of prevalences}

\item{e_time}{Vector of expected times}

\item{v_time}{Vector of variances}
}
\description{
Plot a scatter plot with error bars for time to recruit against prevalence
}
