% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{is.accrual}
\alias{is.accrual}
\title{Check whether an object is of class "accrual".}
\usage{
is.accrual(x, ...)
}
\arguments{
\item{x}{Object to test}

\item{...}{For R CMD check compatibility.}
}
\value{
TRUE or FALSE
}
\description{
Check whether an object is of class "accrual".
}
