% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{is.trial_structure}
\alias{is.trial_structure}
\title{Check whether an object is of class "trial_structure".}
\usage{
is.trial_structure(x, ...)
}
\arguments{
\item{x}{Object to test}

\item{...}{For R CMD check compatibility.}
}
\value{
TRUE or FALSE
}
\description{
Check whether an object is of class "trial_structure".
}
