% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graphics.R
\name{label_vlines}
\alias{label_vlines}
\title{Adds labels for vlines to accrual plots}
\usage{
label_vlines(p, target, target_names, size = 6)
}
\arguments{
\item{p}{Ggplot object.}

\item{target}{Vector of x axis positions of vlines.}

\item{target_names}{Vector of target names (excluding the
word \code{target}).}

\item{size}{Font size (in ggplot measure) for labels;
defaults to 6.}
}
\description{
Adds labels for vlines to accrual plots
}
