% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-utils.R
\name{makeifnot_dir}
\alias{makeifnot_dir}
\title{Create directory if it doesn't already exist, or
issue error if it does exist but the minimum access
requirements are not met, to avoid using directories
created for another purpose.}
\usage{
makeifnot_dir(file_path)
}
\arguments{
\item{file_path}{Location of directory to check/create}
}
\description{
Create directory if it doesn't already exist, or
issue error if it does exist but the minimum access
requirements are not met, to avoid using directories
created for another purpose.
}
