% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-batches.R
\name{matrix_to_long}
\alias{matrix_to_long}
\title{Format batch accrual data in long format for plotting.}
\usage{
matrix_to_long(data)
}
\arguments{
\item{data}{Matrix of accrual data.}
}
\description{
Format batch accrual data in long format for plotting.
}
