% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graphics.R
\name{plot.armaccrual}
\alias{plot.armaccrual}
\title{Plot simulations of recruitment to given arm over time.}
\usage{
\method{plot}{armaccrual}(x, ..., arm_colour, target, target_names, plot_id)
}
\arguments{
\item{x}{Matrix with columns for each simulation.}

\item{...}{For compliance with plot.default().}

\item{target}{Vector of targets for recruitment. First two
should be those directly relevant to the subject of the graph.}

\item{target_names}{Vector of target names, for labelling.}

\item{plot_id}{Type of plot for title, e.g. "Treatment A".}
}
\description{
Plot simulations of recruitment to given arm over time.
}
