% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graphics.R
\name{plot.armtotals}
\alias{plot.armtotals}
\title{Plot distributions of recruitment to arms at given time.}
\usage{
\method{plot}{armtotals}(data, target, target_names)
}
\arguments{
\item{data}{Matrix with columns for each recruitment arm,
including control.}

\item{target}{Vector of targets for recruitment.}

\item{target_names}{Vector of target names, for labelling.}
}
\description{
Plot distributions of recruitment to arms at given time.
}
