% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{plot}
\alias{plot}
\alias{plot.accrual}
\alias{plot.trial_strucutre}
\title{Plot method for an object of class \code{accrual}.  Creates
a line plot of cumulative recruitment, grouped by trial arm,
using ggplot2.}
\arguments{
\item{accrual_obj}{Object of class \code{accrual}.}

\item{plot_prefix}{Prefix for file name to identify plot type.
Defaults to \code{accrual_plot}.}

\item{run_time}{Specify a particular instance of \code{spine()}
execution using a date-time format \code{yyyy-mm-dd-hh-mm-ss}.}

\item{output_path}{= Directory where the output files from the
\code{spine()} instance are located.}

\item{figs_path}{Folder where figures generated during execution
will be stored; defaults to the \code{figures} subdirectory in
\code{output_path}.}

\item{x}{An object of class \code{trial_structure}.}
}
\description{
Plot method for an object of class \code{accrual}.  Creates
a line plot of cumulative recruitment, grouped by trial arm,
using ggplot2.

Prints initial trial structure as a matrix of prevalences by recruitment
and experimental arms.
}
