% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graphics.R
\name{plot.targetweek}
\alias{plot.targetweek}
\title{Plot distribution of time to accrual to a specified
target for a specified arm.}
\usage{
\method{plot}{targetweek}(x, ..., arm_colour, target, target_names = NULL, plot_id)
}
\arguments{
\item{x}{Vector of times to accrual.}

\item{...}{For compliance with plot.default().}

\item{arm_colour}{Hexadecimal colour associated with arm.}

\item{target}{Accrual target, for labelling.}

\item{target_names}{Target name, for labelling.}

\item{plot_id}{Type of plot for title, e.g. "Treatment A".}
}
\description{
Plot distribution of time to accrual to a specified
target for a specified arm.
}
