% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{register-S7-methods}
\alias{register-S7-methods}
\alias{.onLoad}
\title{Ensure methods are registered when methods for generics are implemented in
other packages (S3, S4 & S7). This works like export directives in the
NAMESPACE does for S3 and S4.}
\usage{
.onLoad(...)
}
\arguments{
\item{...}{Parameters passed to S7::methods_register()}
}
\description{
Ensure methods are registered when methods for generics are implemented in
other packages (S3, S4 & S7). This works like export directives in the
NAMESPACE does for S3 and S4.
}
