% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sens_analysis}
\alias{sens_analysis}
\title{Driver for the procedure}
\usage{
sens_analysis(
  target_arm_size = 308,
  site_rates,
  no_centres = NULL,
  centre_starts = NULL,
  fixed_site_rates = TRUE,
  fixed_centre_starts = TRUE,
  site_caps = NULL,
  figs_path = "output_data/figures/"
)
}
\arguments{
\item{target_arm_size}{Number of patients required per treatment arm}

\item{site_rates}{Vector of expected site accrual per month, one per site.
If all site rates are equal, one value can be given, if no_centres is
also specified.}

\item{no_centres}{Number of centres (defaults to the length of site_rates)}

\item{centre_starts}{Vector of site opening months, one per site.
if not specified, defaults to all starting at time zero.}

\item{fixed_site_rates}{Logical; are site rates to be considered as
fixed (TRUE) or do they vary randomly according to the gamma distribuiton
(FALSE); defaults to TRUE.}

\item{fixed_centre_starts}{Logical; do sites open at the beginning of
the month specified (TRUE) or does the time vary randomly; defaults to TRUE.}

\item{site_caps}{Vector of site caps; if all sites are capped to the same
value can be a scalar. Defaults to NULL, for no site capping.}
}
\description{
Driver for the procedure
}
\examples{
sensitivity(10, 100)

}
