% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{set_site_rates}
\alias{set_site_rates}
\title{Generate site rates for each site, drawing from a gamma
distribution if the rates are not fixed.}
\usage{
set_site_rates(mean_rate, fixed_site_rates, var_lambda)
}
\arguments{
\item{mean_rate}{Vector of expected recruitment rates for all sites,
in patients per month (from the centres file).}

\item{fixed_site_rates}{Logical: whether site rates are fixed or
distributed according to the gamma distribution.}

\item{var_lambda}{Variance of the gamma distribution; NULL if
site rates are fixed.}
}
\value{
Vector of site rates in patients per week.
}
\description{
Generate site rates for each site, drawing from a gamma
distribution if the rates are not fixed.
}
