% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{summary}
\alias{summary}
\alias{summary.accrual}
\alias{summary.trial_structure}
\title{Summary of predicted accrual.}
\arguments{
\item{...}{Additional arguments passed to summary().}

\item{object}{An object of class \code{trial_structure}.}

\item{..., }{Additional arguments passed to print().}

\item{digits}{A non-null value for digits specifies
the minimum number of significant digits to be
printed in values. The default, NULL, uses
getOption("digits"). (For the interpretation for
complex numbers see signif.) Non-integer values
will be rounded down, and only values greater than
or equal to 1 and no greater than 22 are accepted.}
}
\description{
Summary of predicted accrual.

Summary of trial structure.
}
