% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-utils.R
\name{target_group}
\alias{target_group}
\title{Group arm names by target value}
\usage{
target_group(target_df, target_col)
}
\arguments{
\item{target_df}{Dataframe with first column containing the name
of the arm and subsequent columns containing recruitment targets
for that arm at different time periods.}

\item{target_col}{Column number for the time period of interest
in \code{target_df}.}
}
\description{
Group arm names by target value
}
