% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-utils.R
\name{targets_tolong}
\alias{targets_tolong}
\title{Convert targets dataframe to long format}
\usage{
targets_tolong(target_df)
}
\arguments{
\item{target_df}{Dataframe of targets in target file or extended format}
}
\description{
Convert targets dataframe to long format
}
