% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graphics.R
\name{theme_bma}
\alias{theme_bma}
\title{Theme for ggplot2}
\usage{
theme_bma(base_size = 10)
}
\arguments{
\item{base_size}{Legend title size, all other sizes scaled appropriately
to this}
}
\description{
Theme for ggplot2
}
