% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graphics.R
\name{threshold_week}
\alias{threshold_week}
\title{Extract from a simulation * week accrual matrix for a
specific trial arm, the week in which an arm target
threshold or thresholds are first met or exceeded for
each simulation.}
\usage{
threshold_week(accrual, targets)
}
\arguments{
\item{accrual}{Matrix of accrual data for a given arm}

\item{targets}{Target threshold(s) for recruitment.}
}
\value{
List of vectors of the weeks in which each simulation
reaches the target accrual threshold.
}
\description{
Extract from a simulation * week accrual matrix for a
specific trial arm, the week in which an arm target
threshold or thresholds are first met or exceeded for
each simulation.
}
