% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{trial_structure}
\alias{trial_structure}
\title{A class for an object to contain the proportions of arriving patients to be
allocated to each arm of the trial.}
\usage{
trial_structure(
  props_df = S7::class_missing,
  arms_ls = S7::class_missing,
  centres_df = S7::class_missing,
  ...
)
}
\description{
A class for an object to contain the proportions of arriving patients to be
allocated to each arm of the trial.
}
\section{Slots}{

\describe{
\item{\code{recruit_arm_prevalence}}{Numeric vector of the expected proportion
of patients eligible for each recruitment arm.}

\item{\code{recruit_arm_names}}{Character vector of the names of the recruitment
arms.}

\item{\code{treatment_arm_ids}}{Named list of lists of recruitment arms by
treatment arm.}

\item{\code{recruit_arm_id}}{Automatically generated integer vector of the ID
numbers of the recruitment arms.}

\item{\code{treatment_counts}}{Automatically generated named integer vector of
number of recruitment arms recruiting to each treatment arm.}

\item{\code{treatment_arm_struct}}{Automatically generated logical matrix of
treatment arms by recruitment arms.}

\item{\code{treatment_arm_prevalence}}{Automatically generated matrix of recruitment
prevalences of treatment arms by recruitment arms}
}}

