% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{week_accrue}
\alias{week_accrue}
\title{Randomise a week's expected accrual amongst the sites, according to
prevalence.}
\usage{
week_accrue(accrual_obj, struct_obj, ...)
}
\arguments{
\item{class_list}{A list of one object of class "accrual" and one
of class "trial_structure".}
}
\value{
Matrix of week's accrual by site and recruitment arm.
}
\description{
Randomise a week's expected accrual amongst the sites, according to
prevalence.
}
