% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{accrual}
\alias{accrual}
\title{A class to contain the accrued patient allocations,
by site, treatment/control arm and week}
\usage{
accrual(treatment_arm_ids, shared_control, fixed_site_rates, target_arm_size, 
target_control, target_interim, accrual_period, interim_period, 
var_lambda, centres_df)
}
\arguments{
\item{treatment_arm_ids}{Named list of lists of recruitment arms by
treatment arm.}

\item{shared_control}{TRUE if all experimental arms share one control arm;
FALSE if each has their own}

\item{fixed_site_rates}{TRUE if expected site rate to be used; FALSE
draws the site rate from a gamma distribution}

\item{target_arm_size}{Number of subjects required for each treatment arm.}

\item{target_control}{Number of subjects required for control arm(s).}

\item{target_interim}{Number of subjects required for treatment arm at
interim analysis.}

\item{accrual_period}{Number of weeks in recruitment period.}

\item{interim_period}{Number of weeks to recruit for interim analysis.}

\item{var_lambda}{Variance of site recruitment rates.}

\item{centres_df}{Dataframe with columns "site", "start_month", "mean_rate",
"region" and "site_cap"}
}
\description{
A class to contain the accrued patient allocations,
by site, treatment/control arm and week
}
\section{Slots}{

\describe{
\item{\code{accrual}}{3-D array with axes site, experimental arm
and week}

\item{\code{target_arm_size}}{Number of subjects required for each treatment arm.}

\item{\code{target_control}}{Number of subjects required for control arm(s).}

\item{\code{target_interim}}{Number of subjects required for treatment arm at
interim analysis.}

\item{\code{accrual_period}}{Number of weeks in recruitment period.}

\item{\code{interim_period}}{Number of weeks to recruit for interim analysis.}

\item{\code{phase_changes}}{Vector of week numbers when arms closed}

\item{\code{site_closures}}{Vector of weeks sites closed; NA indicates open}

\item{\code{week}}{Current recruitment week}

\item{\code{active_arms}}{Vector of indices of open arms}

\item{\code{active_sites}}{Vector of indices of open sites}

\item{\code{shared_control}}{TRUE if a shared control arm is being used, else FALSE}

\item{\code{fixed_site_rates}}{TRUE if expected site rate to be used; FALSE
draws the site rate from a gamma distribution}

\item{\code{site_in_region}}{Vector of indices for which set of expected
prevalences each site should use}

\item{\code{site_cap}}{Vector of maximum number of patients for each site}

\item{\code{site_mean_rate}}{Vector of expected recruitment rates for each site}

\item{\code{site_rate}}{Vector of recruitment rates for each site, drawn from a
gamma distribution}

\item{\code{start_week}}{Vector of weeks that each site opens recruitment}

\item{\code{site_index}}{Vector of index numbers for each site}

\item{\code{treatment_arm_ids}}{Named list of lists of recruitment arms by
treatment arm.}

\item{\code{var_lambda}}{Variance of site recruitment rates.}
}}

