% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graphics.R
\name{accrual_arm_plot}
\alias{accrual_arm_plot}
\title{Plot single arm accrual plot}
\usage{
accrual_arm_plot(data_df, arm_colours, treatment_arms, targets, plot_id, i)
}
\description{
Bodge, fix later
}
