% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{accrue_week}
\alias{accrue_week}
\title{Generate accrual for a week and assign it to the
accrual object. Increments the property "week",
which holds the next week number to accrue.}
\usage{
accrue_week(accrual_obj, struct_obj, ...)
}
\arguments{
\item{accrual_obj}{An object of class "accrual"}

\item{struct_obj}{An object of class "trial_structure"}
}
\value{
An object of class "accrual"
}
\description{
Generate accrual for a week and assign it to the
accrual object. Increments the property "week",
which holds the next week number to accrue.
}
