% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{apply_arm_cap}
\alias{apply_arm_cap}
\title{Implement arm cap on week's accrual to experimental arms}
\usage{
apply_arm_cap(accrual_obj, struct_obj, ...)
}
\arguments{
\item{accrual_obj}{Object of class \code{accrual}}

\item{struct_obj}{Object of class \code{trial_structure}}

\item{...}{For compliance with R CMD check.}
}
\value{
Modified accrual object with capped week's accrual
}
\description{
Implement arm cap on week's accrual to experimental arms
}
