% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-batches.R
\name{batch}
\alias{batch}
\title{Run a number of batches of recruitment prediction and
collect summary statistics on arm closures, and final
recruitment totals for all experimental and control arms}
\usage{
batch(
  n = 100,
  centres_file = "centres.csv",
  arms_file = "arms.json",
  data_path = "extdata/",
  output_path = "../biomkrAccrual_output_data/",
  figs_path = paste0(output_path, "figures/"),
  accrual_period = 6,
  interim_period = 3,
  precision = 10,
  ctrl_ratio = c(1, 1),
  target_arm_size = 60,
  target_interim = target_arm_size/2,
  target_control = 180,
  target_interim_control = target_control/2,
  shared_control = TRUE,
  fixed_centre_starts = TRUE,
  fixed_site_rates = FALSE,
  fixed_region_prevalences = FALSE,
  quietly = TRUE,
  keep_files = FALSE
)
}
\arguments{
\item{n}{Number of instances to run (defaults to 100)}

\item{centres_file}{File containing recruitment centre data
(defaults to "centres.csv")}

\item{arms_file}{File containing list of which recruitment
arms recruit to which experimental arm}

\item{data_path}{Folder where \code{centres_file}, \code{prop_file} and
\code{arms_file} are located. Defaults to the location of the package
example data in the package installation; this should be changed.}

\item{output_path}{Folder where data generated during execution
will be stored; defaults to \verb{../biomkrAccrual_output_data/}.}

\item{figs_path}{Folder where figures generated during execution
will be stored; defaults to the \code{figures} subdirectory in
\code{output_path}.}

\item{accrual_period}{Maximum number of weeks to recruit
(defaults to 80)}

\item{target_arm_size}{Maximum size for all experimental arms
(defaults to 308)}

\item{target_interim}{Recruitment target for experimental arms
at interim analysis; defaults to \code{target_arm_size / 2}.}

\item{target_control}{Maximum size for all control arms
(defaults to 308)}

\item{target_interim_control}{Recruitment target for control arms
at interim analysis; defaults to \code{target_control / 2}.}

\item{shared_control}{= TRUE,}

\item{fixed_centre_starts}{TRUE if centres are assumed to start
exactly when planned; FALSE if some randomisation should be added.}

\item{fixed_site_rates}{TRUE if centre recruitment rates should
be treated as exact; FALSE if they should be drawn from a gamma
distribution with a mean of the specified rate.}

\item{fixed_region_prevalences}{TRUE if biomarker prevalences
should be considered to be identical for all sites within a
region; FALSE if they should be drawn from a Dirichlet distribution
with a mean of the specified prevalence.}

\item{quietly}{If TRUE, do not display information and plots from
individual runs within the batch. Defaults to TRUE.}

\item{keep_files}{If FALSE, do not save data or plots from individual
runs within the batch.  Defaults to FALSE.}
}
\value{
Dataframe of site closing times

Dataframe of experimental arm totals
}
\description{
Run a number of batches of recruitment prediction and
collect summary statistics on arm closures, and final
recruitment totals for all experimental and control arms
}
