% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_batches.R
\name{batch}
\alias{batch}
\title{Run a number of batches of recruitment prediction and
collect summary statistics on arm closures, and final
recruitment totals for all experimental and control arms}
\usage{
batch(
  n = 100,
  centres_file = "centres.csv",
  arms_file = "arms.json",
  data_path = "inst/extdata/",
  output_path = "output_data/",
  accrual_period = 10,
  target_arm_size = 100,
  shared_control = TRUE,
  fixed_site_rates = FALSE
)
}
\arguments{
\item{n}{Number of instances to run (defaults to 100)}

\item{centres_file}{File containing recruitment centre data
(defaults to "centres.csv")}

\item{arms_file}{File containing list of which recruitment
arms recruit to which experimental arm}

\item{accrual_period}{Maximum number of weeks to recruit
(defaults to 80)}

\item{target_arm_size}{Maximum size for all experimental arms
(defaults to 308)}

\item{shared_control}{= TRUE,}
}
\value{
Dataframe of site closing times

Dataframe of experimental arm totals
}
\description{
Run a number of batches of recruitment prediction and
collect summary statistics on arm closures, and final
recruitment totals for all experimental and control arms
}
