% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{do_dirichlet_draws}
\alias{do_dirichlet_draws}
\title{Draws from dirichlet regression model for biomarker prevalences
to create the prevalence matrix.}
\usage{
do_dirichlet_draws(region_prevalence, sites_in_region, precision)
}
\arguments{
\item{region_prevalence}{Dataframe with one column for each
region and one row for each biomarker, containing prevalences as
probabilities.}

\item{sites_in_region}{Vector with a region index number for each
site, the index determined by the order of the columns in
\code{region_prevalence}.}

\item{precision}{Variability decreases as precision increases.}
}
\description{
Draws from dirichlet regression model for biomarker prevalences
to create the prevalence matrix.
}
