% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{do_sensitivity_plot_simultaneous}
\alias{do_sensitivity_plot_simultaneous}
\title{Creates sensitivity plot for Poisson distribution with
specified site rates for simultaneous start}
\usage{
do_sensitivity_plot_simultaneous(
  target_arm_size,
  site_rates,
  plot_name,
  figs_path
)
}
\arguments{
\item{target_arm_size}{Number of patients to be recruited.}

\item{site_rates}{Total site rate.}

\item{plot_name}{Name for sensitivity plot file (.png will be appended).}
}
\description{
Creates sensitivity plot for Poisson distribution with
specified site rates for simultaneous start
}
