% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graphics.R
\name{plot.accrualplotdata}
\alias{plot.accrualplotdata}
\title{S3 method to plot predicted recruitment from a long format
dataframe of class "accrualplotdata".}
\usage{
\method{plot}{accrualplotdata}(
  data,
  plot_prefix = "accrual_plot",
  run_time = NULL,
  output_path = "../biomkrAccrual_output_data/",
  figs_path = paste0(output_path, "figures/")
)
}
\arguments{
\item{data}{long format dataframe with columns "Week",
"Arm" and "Recruitment".}

\item{plot_prefix}{Prefix for file name to identify plot type.
Defaults to \code{accrual_plot}.}

\item{run_time}{Specify a particular instance of \code{spine()}
execution using a date-time format \code{yyyy-mm-dd-hh-mm-ss}.}

\item{output_path}{= Directory where the output files from the
\code{spine()} instance are located.}

\item{figs_path}{Folder where figures generated during execution
will be stored; defaults to the \code{figures} subdirectory in
\code{output_path}.}
}
\description{
S3 method to plot predicted recruitment from a long format
dataframe of class "accrualplotdata".
}
