% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.R
\name{rdirichlet_alt}
\alias{rdirichlet_alt}
\title{Generate a matrix of \code{n} rows of sets of probabilities
generated from the Dirichlet distribution, each row
summing to 1. Uses the alternative \eqn{\mu} \eqn{\phi}
parameterisation for the Dirichlet distribution,
representing means and precision.}
\usage{
rdirichlet_alt(n = 1, mu = c(0.3, 0.3, 0.3), phi = 10)
}
\arguments{
\item{n}{Number of sets of probabilities (defaults to 1)}

\item{mu}{Vector of mean values for each probability in the set
(defaults to c(0.001, 0.029. 0.7)). Must be greater than 0 and
finite, and contain at least two values.}

\item{phi}{Parameter representing precision, where precision is
1/variance. Must be positive and finite. Defaults to 10.}
}
\description{
Generate a matrix of \code{n} rows of sets of probabilities
generated from the Dirichlet distribution, each row
summing to 1. Uses the alternative \eqn{\mu} \eqn{\phi}
parameterisation for the Dirichlet distribution,
representing means and precision.
}
\examples{
rdirichlet_alt(n = 3, mu = c(0.001, 0.029, 0.7), phi = 10)

}
