% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{set_site_rates}
\alias{set_site_rates}
\title{Method to increment site rates by gamma-distributed rates of sites
opening an accrual pathway in the current week.}
\usage{
set_site_rates(obj, ...)
}
\arguments{
\item{obj}{An object of type "accrual".}

\item{...}{For R CMD check compatibility.}
}
\value{
Modified object with new site rates
}
\description{
Method to increment site rates by gamma-distributed rates of sites
opening an accrual pathway in the current week.
}
