% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accrual.R
\name{site_sums}
\alias{site_sums}
\title{Sum accrual array by site, to enable checking against site caps}
\usage{
site_sums(obj, ...)
}
\arguments{
\item{obj}{Object of class "accrual"}
}
\value{
vector of total accrual by recruitment site
}
\description{
Sum accrual array by site, to enable checking against site caps
}
