% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spine.R
\name{spine}
\alias{spine}
\title{Driver for the procedure}
\usage{
spine(
  target_arm_size = 308,
  target_interim = target_arm_size/2,
  target_control = 704,
  accrual_period = 36,
  shared_control = TRUE,
  ctrl_ratio = c(1, 1),
  no_samples = 100,
  centres_file = "centres.csv",
  prop_file = "proportions.csv",
  centre_start_file = "centre_starts.csv",
  arms_file = "arms.json",
  phase_file = "phase_change_weeks.csv",
  average_file = "mean_recruitment.csv",
  site_rate_file = "site_rates.csv",
  data_path = "inst/extdata/",
  output_path = "biomkrAccrual_output_data/",
  figs_path = paste0(output_path, "figures/"),
  fixed_centre_starts = TRUE,
  fixed_site_rates = FALSE
)
}
\arguments{
\item{target_arm_size}{Number of patients required per treatment arm}
}
\description{
Driver for the procedure
}
\examples{
spine()

}
