% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{summary}
\alias{summary}
\alias{summary.accrual}
\alias{summary.trial_structure}
\title{Summary of predicted accrual.}
\arguments{
\item{x}{An object of class \code{trial_structure}.}
}
\description{
Summary of predicted accrual.

Summary of trial structure.
}
