% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{treat_sums}
\alias{treat_sums}
\alias{treat_sums.biomkrAccrual::accrual}
\title{Sum accrual array element of accrual object, by experimental
arm (including control).}
\usage{
treat_sums(x, control_total, ...)

\method{treat_sums}{`biomkrAccrual::accrual`}(x, control_total = FALSE)
}
\arguments{
\item{x}{Object of class \code{accrual}.}

\item{control_total}{Logical; if TRUE return single total for all
control arms}

\item{...}{When called on an array the arguments no_treat_arms and
shared_control may be necessary. Additional arguments are
unnecessary if calling treat_sums.biomkrAccrual::accrual().}
}
\value{
vector of total accrual by experimental arm
}
\description{
Sum accrual array element of accrual object, by experimental
arm (including control).
}
