% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accrual.R
\name{treat_sums}
\alias{treat_sums}
\title{Sum accrual array by experimental arm (including control)}
\usage{
treat_sums(obj, ...)
}
\arguments{
\item{obj}{Object of class "accrual"}

\item{control_total}{Logical; if TRUE return single total for control arms}
}
\value{
vector of total accrual by experimental arm
}
\description{
Sum accrual array by experimental arm (including control)
}
