% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{trial-structure}
\alias{trial-structure}
\alias{trial_structure}
\alias{remove_treat_arms}
\title{A class for an object to contain the proportions of arriving patients to be
allocated to each arm of the trial.}
\usage{
trial_structure(props_df = S7::class_missing, arms_ls = S7::class_missing, ...)

remove_treat_arms(obj, ...)
}
\arguments{
\item{arms}{vector or scalar of integer arm ID numbers}
}
\description{
A class for an object to contain the proportions of arriving patients to be
allocated to each arm of the trial.

Close off treatment arms.
In the list of treatment arm IDs, replaces the vector of recruitment
arm IDs with NA.
If any recruitment arms are closed, sets their prevalence to zero and
recalculates prevalence vector, on the assumption that no patients with
those characteristics will be recruited from that point.
Class object will automatically generate new trial structure and
prevalence matrices.
}
\section{Slots}{

\describe{
\item{\code{recruit_arm_prevalence}}{Numeric vector of the expected proportion
of patients eligible for each recruitment arm.}

\item{\code{recruit_arm_names}}{Character vector of the names of the recruitment
arms.}

\item{\code{treatment_arm_ids}}{Named list of lists of recruitment arms by
treatment arm.}

\item{\code{recruit_arm_id}}{Automatically generated integer vector of the ID
numbers of the recruitment arms.}

\item{\code{treatment_counts}}{Automatically generated named integer vector of
number of recruitment arms recruiting to each treatment arm.}

\item{\code{treatment_arm_struct}}{Automatically generated logical matrix of
treatment arms by recruitment arms.}

\item{\code{treatment_arm_prev}}{Automatically generated matrix of recruitment
prevalences of treatment arms by recruitment arms}
}}

