% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{remove_treat_arms}
\alias{remove_treat_arms}
\title{Close off treatment arms.
In the list of treatment arm IDs, replaces the vector of recruitment
arm IDs with NA.
If any recruitment arms are closed, sets their prevalence to zero and
recalculates prevalence vector, on the assumption that no patients with
those characteristics will be recruited from that point.
Class object will automatically generate new trial structure and
prevalence matrices.}
\usage{
remove_treat_arms(obj, ...)
}
\arguments{
\item{arms}{vector or scalar of integer arm ID numbers}
}
\description{
Close off treatment arms.
In the list of treatment arm IDs, replaces the vector of recruitment
arm IDs with NA.
If any recruitment arms are closed, sets their prevalence to zero and
recalculates prevalence vector, on the assumption that no patients with
those characteristics will be recruited from that point.
Class object will automatically generate new trial structure and
prevalence matrices.
}
