% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accrual.R
\name{week_accrue}
\alias{week_accrue}
\title{Randomise a week's expected accrual amongst the sites, according to
prevalence.}
\usage{
week_accrue(obj, ...)
}
\arguments{
\item{obj}{Accrual object}
}
\value{
Matrix of week's accrual by site and recruitment arm
}
\description{
Randomise a week's expected accrual amongst the sites, according to
prevalence.
}
