% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{week_accrue}
\alias{week_accrue}
\title{Randomise a week's expected accrual amongst the sites, according to
prevalence.}
\usage{
week_accrue(accrual_obj, struct_obj, ...)
}
\arguments{
\item{accrual_obj}{An object of class \code{accrual}.}

\item{struct_obj}{An object of class \code{trial_structure}.
be treated as exact; FALSE if they should be drawn from a gamma
distribution with a mean of the specified rate.}
}
\value{
Matrix of week's accrual by site and recruitment arm.
}
\description{
Randomise a week's expected accrual amongst the sites, according to
prevalence.
}
